+++ src/sailpoint/report.py
@@ -12,6 +12,7 @@ class IDNReport:
     def __init__(self, api, idn):
         self.api = api
         self.idn = idn  # the idn util object
+        self.user_cache = {}  # User cache index for get_user_by_id_cache fn
 
     def get_disabled_gg_members(self):
         disabled_gg_members = []
@@ -41,6 +42,19 @@ class IDNReport:
 
                         disabled_gg_members.append(record)
         return disabled_gg_members
+
+    # Build cache so we don't look up users multiple times
+    def get_user_by_id_cache(self, user_id):
+        '''
+        Gets a user by their ID, but also caches results so next time the
+        lookup doesn't have to go to the API
+        '''
+        if user_id in self.user_cache:
+            return self.user_cache[user_id]
+        else:
+            user = self.idn.get_user_by_id(user_id)
+            self.user_cache[user_id] = user
+            return user
 
     def get_disabled_ap_owners(self):
         # It would be much faster to get the list of disabled AP owners using
@@ -88,6 +102,60 @@ class IDNReport:
 
                     disabled_ap_owners.append(record)
         return disabled_ap_owners
+
+    def gg_membership(self):
+        '''
+        Outputs a list of Governance groups, members and the associations one per line
+        '''
+        log.warning('This function takes a while to run')
+        gg_members = {}
+        # Retrieve all Governance Groups with their members and connections
+        gg_list = self.idn.get_gg(members=True, connections=True)
+        count = 0
+
+        # Iterate over each Governance Group
+        for gg in gg_list:
+            count += 1
+            log.debug(count)
+            log.debug(gg.get('name'))
+            gg_members_list = gg.get('members', [])
+
+            # Iterate over all members in the current Governance Group
+            for member in gg_members_list:
+                # Log the member information
+                # log.info(member)
+                # Retrieve details for each member
+                user = self.get_user_by_id_cache(member.get('id'))
+                attributes = user.get('attributes', {})
+
+                # Iterate over all connections for the Governance Group
+                for connection in gg.get('connections', []):
+                    # Construct the record for the current member
+                    record = {
+                        'Governance Group': gg.get('name'),
+                        'Member Name': user.get('name'),
+                        'Identity ID': user.get('id'),
+                        'LifeCycle': attributes.get('cloudLifecycleState'),
+                        'Employee Number': user.get('employeeNumber'),
+                        'AD Username': attributes.get('username'),
+                        'Item Type': connection.get('object', {}).get(
+                            'type', ''
+                        ),
+                        'Item Name': connection.get('object', {}).get(
+                            'name', ''
+                        ),
+                        'Item Description': connection.get('object', {}).get(
+                            'description', ''
+                        ),
+                        'Item ID': connection.get('object', {}).get('id', ''),
+                    }
+                    log.debug(pretty_repr(record))
+                    if not user.get('id') in gg_members:
+                        gg_members[user.get('id')] = []
+                    gg_members[user.get('id')].append(record)
+                    # log.info(pretty(gg_members))
+
+        return gg_members
 
     def get_ai_recommendations(self, id):
         '''